Create newmeal() function
This function checks to see if there is a directory called 'recipe_book' in the cwd, and creates one if it's not present. It then allows a user to input a new meal by listing the number of ingredients and their amounts. Then a file is written with the title, the number of servings and the ingredient list and amounts including units.