Create menu_and_list
All functions are now in one script together with the main process that runs the individual functions. Not all defined functions have a docstring and more annotations are needed to describe the process. Currently on version 0.1. To go to version 0.2, all defined functions must have a docstring and all sections of the code must be annotated.